+++ cmd/vegawallet/commands/key_describe.go
@@ -30,7 +30,7 @@ type DescribeKeyHandler func(api.AdminDescribeKeyParams) (api.AdminDescribeKeyRe
 
 func NewCmdDescribeKey(w io.Writer, rf *RootFlags) *cobra.Command {
 	h := func(params api.AdminDescribeKeyParams) (api.AdminDescribeKeyResult, error) {
-		walletStore, err := wallets.InitialiseStore(rf.Home)
+		walletStore, err := wallets.InitialiseStore(rf.Home, false)
 		if err != nil {
 			return api.AdminDescribeKeyResult{}, fmt.Errorf("couldn't initialise wallets store: %w", err)
 		}