+++ cmd/vegawallet/commands/permissions_revoke.go
@@ -35,7 +35,7 @@ type RevokePermissionsHandler func(api.AdminRevokePermissionsParams) error
 
 func NewCmdRevokePermissions(w io.Writer, rf *RootFlags) *cobra.Command {
 	h := func(params api.AdminRevokePermissionsParams) error {
-		walletStore, err := wallets.InitialiseStore(rf.Home)
+		walletStore, err := wallets.InitialiseStore(rf.Home, false)
 		if err != nil {
 			return fmt.Errorf("couldn't initialise wallets store: %w", err)
 		}